crypto: fix incorrect MAC comment about SUF-CMA security
All variants of HMAC are SUF-CMA secure. In short, a MAC is SUF-CMA
secure if it is UF-CMA secure *and* an adversary cannot forge
a *different* tag that also authenticates for the same key and message.

This is different from hmac(k, msg) where |k| > B since since in that
case the tags are the exact same.

Signed-off-by: Eric Lagergren <elagergren@spideroak.com>