crypto: `Mac` should not require fixed-size keys (#34)
Add `Mac::try_new` to support variable length keys. We still keep
`Mac::new`, `Mac::Key`, etc. around to support keys stored in HSMs.

I'm not a huge fan of the API, but it's difficult to balance the
following:

1. Some MACs only use fixed-length keys (CMAC, ...)
2. Some MACs accept variable length keys (HMAC, KMAC, ...)
3. We need a `type Key` associated type to support keys stored in an
HSM.

Ideally, we'd use something like this:

```rust
trait Mac {
    type Key;
}
impl Mac for Kmac {
    type Key = KmacKey;
}
struct KmacKey<'a>(&'a [u8]);
```

But we're missing the lifetime for `type Key = KmacKey`: either
`Mac::Key` needs a lifetime or `Kmac` needs a lifetime. The former
complicates nearly every use of `Mac::Key` and the latter is simply
wrong: `Kmac`'s key only needs to live as long as the call to `<Kmac as
Mac>::new`.

It's also complicated by the `for<'a> Import<&'a [u8]>` bound on
`SecretKey`. That bound requires the `&[u8]` be valid for *any*
lifetime, whereas

```rust
impl<'a> Import<&'a [u8]> for KmacKey<'a> { ... }
```

requires the `&[u8]` be valid for only `'a`. In other words, the impl
isn't general enough to for `KmacKey` to implement `SecretKey`. (We also
can't use `struct KmacKey(Vec<u8>)` because we're no-alloc by default.)

I considered removing the `for<'a> Import<&'a [u8]>` bound on
`SecretKey`, but it's a particularly invasive change. For example,
`hpke::SealCtx::new` requires importing a `SecretKey`. That then bubbles
up to other APIs, which will eventually affect `aranya-crypto`.

## Misc

While we're here:

- Move the block size stuff out of `Hash`.
- Clippy was complaining about `<ed25519::SigningKey as
ConstantTimeEq>::ct_eq`. Not sure why. It no longer makes a copy of the
secret key, which is nice.
- Get rid of `SecretKey::new` in favor of `csprng::Random`.

Fixes #31

---------

Signed-off-by: Eric Lagergren <elagergren@spideroak.com>